Correct the package name
