+++ src/main/java/org/vanilladb/comm/protocols/floodingcons/FloodingConsensusLayer.java
@@ -1,4 +1,4 @@
-package org.vanilladb.comm.protocols.floodingconsensus;
+package org.vanilladb.comm.protocols.floodingcons;
 
 import org.vanilladb.comm.protocols.events.ProcessListInit;
 import org.vanilladb.comm.protocols.tcpfd.AllProcessesReady;
