+++ src/main/java/org/vanilladb/comm/protocols/floodingcons/Propose.java
@@ -1,4 +1,6 @@
-package org.vanilladb.comm.protocols.floodingconsensus;
+package org.vanilladb.comm.protocols.floodingcons;
+
+import java.util.Set;
 
 import org.vanilladb.comm.protocols.beb.Broadcast;
 
@@ -7,40 +9,51 @@ import net.sf.appia.core.Channel;
 import net.sf.appia.core.Direction;
 import net.sf.appia.core.Session;
 
-public class Decide extends Broadcast {
+public class Propose extends Broadcast {
 	
 	private boolean isInitailized;
 	
-	private Value value;
+	private int roundId;
+	private Set<Value> proposal;
 	
 	// We must provide a public constructor for TcpCompleteSession
 	// in order to reconstruct this on the other side
-	public Decide() {
+	public Propose() {
 		super();
 		this.isInitailized = false;
 	}
 	
-	public Decide(Channel channel, Session source, Value value)
+	public Propose(Channel channel, Session source,
+			int roundId, Set<Value> proposal)
 			throws AppiaEventException {
 		super(channel, Direction.DOWN, source);
-		this.value = value;
+		this.roundId = roundId;
+		this.proposal = proposal;
 		this.isInitailized = true;
 		
 		// Push the data to the message buffer in order to send
 		// through network
-		getMessage().pushObject(value);
+		getMessage().pushInt(roundId);
+		getMessage().pushObject(proposal);
 	}
 	
-	public Value getValue() {
+	public int getRoundId() {
 		if (!isInitailized)
 			recoverData();
-		return value;
+		return roundId;
+	}
+	
+	public Set<Value> getProposal() {
+		if (!isInitailized)
+			recoverData();
+		return proposal;
 	}
 	
 	private void recoverData() {
 		// The data must be recovered from the message buffer
 		// after it is sent through the network.
-		value = (Value) getMessage().popObject();
+		proposal = (Set<Value>) getMessage().popObject();
+		roundId = getMessage().popInt();
 		isInitailized = true;
 	}
 }
