+++ src/main/java/org/vanilladb/comm/protocols/floodingcons/Value.java
@@ -1,4 +1,4 @@
-package org.vanilladb.comm.protocols.floodingconsensus;
+package org.vanilladb.comm.protocols.floodingcons;
 
 import java.io.Serializable;
