+++ src/main/java/org/vanilladb/comm/protocols/tob/TotalOrderBroadcastSession.java
@@ -13,8 +13,8 @@ import java.util.logging.Logger;
 import org.vanilladb.comm.process.ProcessList;
 import org.vanilladb.comm.process.ProcessState;
 import org.vanilladb.comm.protocols.events.ProcessListInit;
-import org.vanilladb.comm.protocols.floodingconsensus.ConsensusRequest;
-import org.vanilladb.comm.protocols.floodingconsensus.ConsensusResult;
+import org.vanilladb.comm.protocols.floodingcons.ConsensusRequest;
+import org.vanilladb.comm.protocols.floodingcons.ConsensusResult;
 import org.vanilladb.comm.protocols.rb.MessageId;
 import org.vanilladb.comm.protocols.tcpfd.AllProcessesReady;
 import org.vanilladb.comm.protocols.tcpfd.FailureDetected;
